Update Localizable.strings (#742)
In the English definition, "lyrics" is used in the plural. In the Spanish translation, there are instances where "lyric" is used in the singular. This change is made for consistency. It also corrects some words that begin with capital letters when it is not correct.

Use double quotation marks to emphasize phrases. Do not use single quotation marks or other types of quotation marks.

Add the missing "lrclib_api" translation.